refactor: split `templates` into modules
Cleaned up the `templates/mod.rs` that contained
a mix of `x86_64` and `aarch64` modules with code for both
cpu templates and creation of cpu configurations.

Moved cpu configuration related code into
`x86_64` and `aarch64` modules in `guest_config`.

Moved cpu template related code into
`x86_64` and `aarch64` modules in `guest_config/templates`.

Updated related modules.
