+++ src/static_templates/src/t2cl.rs
@@ -5,10 +5,10 @@ use vmm::guest_config::cpuid::KvmCpuidFlags;
 use vmm::guest_config::templates::x86_64::{
     CpuidLeafModifier, CpuidRegister, CpuidRegisterModifier, RegisterModifier, RegisterValueFilter,
 };
-use vmm::guest_config::templates::CpuTemplate;
+use vmm::guest_config::templates::CustomCpuTemplate;
 
-pub fn t2cl() -> CpuTemplate {
-    CpuTemplate {
+pub fn t2cl() -> CustomCpuTemplate {
+    CustomCpuTemplate {
         cpuid_modifiers: vec![
             CpuidLeafModifier {
                 leaf: 0x1,
