+++ src/static_templates/tests/main_branch/intel/c3.rs
@@ -5,12 +5,12 @@ use vmm::guest_config::cpuid::KvmCpuidFlags;
 use vmm::guest_config::templates::x86_64::{
     CpuidLeafModifier, CpuidRegister, CpuidRegisterModifier, RegisterValueFilter,
 };
-use vmm::guest_config::templates::CpuTemplate;
+use vmm::guest_config::templates::CustomCpuTemplate;
 
 use crate::main_branch::bit_helper::{BitHelper, BitRangeExt};
 
-pub fn c3() -> CpuTemplate {
-    CpuTemplate {
+pub fn c3() -> CustomCpuTemplate {
+    CustomCpuTemplate {
         cpuid_modifiers: vec![
             leaf_0x1_subleaf_0x0(),
             leaf_0x7_subleaf_0x0(),
